+++ src/index.js
@@ -40,9 +40,21 @@ var suitePath	= suites[0],
 suite.getRunner()
 	 .then(function(runner) {
 	 	return runner.test();
+	 }, function(err) {
+	 	var logger = require('winston').loggers.get('load');
+	 	logger.debug(err.stack);
+
+	 	var errorDescription = ERRORS_LIST[err && err.code];
+	 	if (errorDescription) {
+	 		logger.error(errorDescription.title);
+	 		logger.info(errorDescription.help);
+	 	}
+	 	throw err;
 	 }).fail(function(err) {
-	 	var error = ERRORS_LIST[err && err.code] || { code: 1 };
-	 	statusCode = error.code;
+	 	var error = ERRORS_LIST[err && err.code] || { exitCode: 1 };
+	 	statusCode = error.exitCode;
+	 	if (err.stack)	// TODO: improve detection of what is an actual uncaught exception (We currently have an ambiguity: is the promise rejected because the test failed or because an error occurred? If error, we should tell the user. This is (badly) approximated by the reason for rejection having a stack trace or not.)
+	 		console.error(err.stack);
 	 }).done();	// ensure any uncaught exception gets rethrown